ec2, ec2/ec2test: add InstanceBlockDeviceMapping
This PR introduces the InstanceBlockDeviceMapping data type
and adds the BlockDeviceMappings field to the Instance data
type, enabling users to query instances' block device mappings.

The test server is updated to create volumes for each non-virtual
block device mapping. The volumes are not yet backed by any model,
so attempts to query those resources will fail that is implemented.