Revert signature back to v3.
Some users of goamz will not be able to pass in an io.ReadSeeker. Reverted the signature of PutReaderWithHeader to take in an io.Reader. The reader is then converted to an io.ReadSeeker through either a type assertion, or by reading the entire contents into memory. This allows for an optimization to those who pass in an io.ReadSeeker, but allows use for everyone.